Revert "Database test"

+++ backend/internal/models/dto/profile.go
@@ -3,18 +3,18 @@ package dto
 type ProfileDTO struct {
 	Name     string `json:"name"`
 	LastName string `json:"last_name"`
-	Age      int16  `json:"age"`
-	Id       int32  `json:"id"`
+	Age      int    `json:"age"`
+	Id       int    `json:"id"`
 }
 
 type NewProfileDTO struct {
 	Name     string `json:"name" binding:"required"`
 	LastName string `json:"last_name" binding:"required"`
-	Age      int16  `json:"age" binding:"required,min=18,max=100"`
+	Age      int    `json:"age" binding:"required,min=18,max=100"`
 }
 
 type PatchProfileDTO struct {
 	Name     *string `json:"name"`
 	LastName *string `json:"last_name"`
-	Age      *int16  `json:"age" binding:"omitempty,min=18,max=100"`
+	Age      *int    `json:"age" binding:"omitempty,min=18,max=100"`
 }
